// deleghebowcl/src/app/model/delegaServizio.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/


export interface DelegaServizio {
    uuid?: string;
    codice_servizio?: string;
    data_inizio_delega?: string;
    data_fine_delega?: string;
    data_revoca_delega?: string;
    data_rinuncia_delega?: string;
    description_servizio?: string;
    grado_delega?: string;
    /**
     * i possibili valori dello stato delega modificabili sono  ATTIVA, REVOCATA, RIFIUTATA. Altri valori restituibili dalla business ma non modificabili sono IN SCADENZA e SCADUTA.
     */
    stato_delega?: string;
}